# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gym Multi-Tenant API",
        "description": "API documentation for the Laravel Gym Management system",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/features": {
            "get": {
                "tags": [
                    "Features"
                ],
                "summary": "List all features",
                "operationId": "8db7ce443f5f1605c5052575cdc49c95",
                "responses": {
                    "200": {
                        "description": "List of features"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Features"
                ],
                "summary": "Create a new feature",
                "operationId": "139a69b3e5eea834535440cd43975fcf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "key"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "key": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Feature created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/features/{id}": {
            "get": {
                "tags": [
                    "Features"
                ],
                "summary": "Get a feature by ID",
                "operationId": "facc4fb44f03a604dff28af74e68350a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Feature details"
                    },
                    "404": {
                        "description": "Feature not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Features"
                ],
                "summary": "Update a feature",
                "operationId": "5e0a61bda852feb9d3a3b87c8d005981",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "key": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Feature updated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Features"
                ],
                "summary": "Delete a feature",
                "operationId": "7a4dc18aebc0d9e7931755fa539fb59b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Feature deleted"
                    },
                    "404": {
                        "description": "Feature not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/gyms": {
            "get": {
                "tags": [
                    "Gyms"
                ],
                "summary": "Get list of gyms",
                "operationId": "b6f8616911aeb2fa518f0dab7442fbe3",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Gyms"
                ],
                "summary": "Create new gym with sub-data",
                "operationId": "9f96cc1c72afbb7c741e1cbad2b9e61e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreGymRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/v1/gyms/{id}": {
            "get": {
                "tags": [
                    "Gyms"
                ],
                "summary": "Show a single gym",
                "operationId": "e186626d4d4417a9bc837ac703bad2fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Gyms"
                ],
                "summary": "Update a gym and its related data",
                "operationId": "a1004bba61453841cf4c8fbea22bf27a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateGymRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Gyms"
                ],
                "summary": "Delete a gym and all related records",
                "operationId": "19cad5a9d9b8b2c96c54a8085ce8120e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreGymRequest": {
                "required": [
                    "en_name",
                    "ar_name",
                    "email"
                ],
                "properties": {
                    "en_name": {
                        "description": "StoreGymRequest",
                        "type": "string",
                        "example": "Gold Gym"
                    },
                    "ar_name": {
                        "type": "string",
                        "example": "جولد جيم"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "gold@gym.com"
                    },
                    "mobile": {
                        "type": "string",
                        "example": "+20123456789"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive",
                            "expired",
                            "suspended"
                        ]
                    },
                    "subscription_plan_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "branding": {
                        "properties": {
                            "main_color": {
                                "type": "string",
                                "example": "#111111"
                            },
                            "second_color": {
                                "type": "string",
                                "example": "#eeeeee"
                            },
                            "cover": {
                                "type": "string",
                                "format": "binary"
                            },
                            "logo": {
                                "type": "string",
                                "format": "binary"
                            }
                        },
                        "type": "object"
                    },
                    "policies": {
                        "properties": {
                            "en_terms": {
                                "type": "string",
                                "example": "English terms..."
                            },
                            "ar_terms": {
                                "type": "string",
                                "example": "الشروط بالعربية..."
                            },
                            "en_policy": {
                                "type": "string",
                                "example": "English policy..."
                            },
                            "ar_policy": {
                                "type": "string",
                                "example": "السياسة بالعربية..."
                            },
                            "privacy_file": {
                                "description": "Upload a privacy policy file",
                                "type": "string",
                                "format": "binary"
                            },
                            "side_effects_file": {
                                "description": "Upload a side effects file",
                                "type": "string",
                                "format": "binary"
                            },
                            "faq_file": {
                                "description": "Upload an FAQ document",
                                "type": "string",
                                "format": "binary"
                            }
                        },
                        "type": "object"
                    },
                    "domain": {
                        "required": [
                            "domain"
                        ],
                        "properties": {
                            "domain": {
                                "type": "string",
                                "example": "example.com"
                            },
                            "is_primary": {
                                "type": "boolean",
                                "example": true
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "UpdateGymRequest": {
                "properties": {
                    "en_name": {
                        "type": "string",
                        "example": "Gold Gym"
                    },
                    "ar_name": {
                        "type": "string",
                        "example": "جولد جيم"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "info@gym.com"
                    },
                    "mobile": {
                        "type": "string",
                        "example": "+20123456789"
                    },
                    "subscription_plan_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "branding": {
                        "properties": {
                            "main_color": {
                                "type": "string",
                                "example": "#f4b400"
                            }
                        },
                        "type": "object"
                    },
                    "contacts": {
                        "properties": {
                            "facebook": {
                                "type": "string",
                                "example": "fb.com/updated"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}